+++ src/agent/askSafety.ts
@@ -21,14 +21,6 @@ export function classifyAskQuestionIntent(question: string): AskQuestionIntent {
   return "code";
 }
 
-export function wrapUntrustedBlock(label: string, text: string): string {
-  return [`<${label} untrusted="true">`, text.trim(), `</${label}>`].join("\n");
-}
-
-export function wrapTrustedContext(lines: string[]): string {
-  return ['<context trusted="server">', ...lines, "</context>"].join("\n");
-}
-
 export type AskToolScope = {
   readonly owner: string;
   readonly repo: string;
